Line Shader Implemented & Tested
TO DOs finished in the following files: line.frag, shader.hpp, shader,cpp & fullscreen_triangle_state.hpp.
I have added comments as instructed in the project document but feel free to add more or modify them.
Note: I added a delay in the compare.sh to be able to read the results of the comparison.
Minor Note: All the finished TODOs have been changed to DONE to facilitate looking for unfinished TODOs.