Fixed a minor bug in the Diamond Shader
Tip: In the terminal, write .\run-all.sh to try all the configurations. After it finishes, write .\compare.sh to compare all the results produced from the previous run to the expected output. A window appears and tells you which are correct and which are wrong. Also, check the images in the error folder. If the image corresponding to a certain configuration is black then it's correct. If it has any colored parts, they represent the wrongly drawn parts of the image.